Refactor code: move type definitions to separate file
